Move debug tools to Storybook
- Fix args for Coalescence story
